[chore] #103 Remove HomeBanner component file and Move code to HomeScreen
